Restructure project folders for clean pattern
