+++ pagina_apoyo/README.md
@@ -1,12 +1,23 @@
-# React + Vite
+# Apoyo Logistico - Sitio Informativo
 
-This template provides a minimal setup to get React working in Vite with HMR and some ESLint rules.
+Este proyecto utiliza **React** y **Vite** para construir un sitio orientado al diseño. La estructura de carpetas se ha organizado siguiendo un patrón limpio, separando los componentes, páginas, estilos y recursos estáticos.
 
-Currently, two official plugins are available:
+```
+pagina_apoyo/
+├── public/
+│   └── models/      # modelos 3D para las vistas
+├── src/
+│   ├── assets/      # imágenes y archivos estáticos que se empaquetan
+│   ├── components/  # componentes reutilizables
+│   ├── pages/       # páginas principales del sitio
+│   ├── styles/      # archivos CSS
+│   └── main.jsx     # punto de entrada de la aplicación
+```
 
-- [@vitejs/plugin-react](https://github.com/vitejs/vite-plugin-react/blob/main/packages/plugin-react) uses [Babel](https://babeljs.io/) for Fast Refresh
-- [@vitejs/plugin-react-swc](https://github.com/vitejs/vite-plugin-react/blob/main/packages/plugin-react-swc) uses [SWC](https://swc.rs/) for Fast Refresh
+## Scripts básicos
 
-## Expanding the ESLint configuration
+- `npm run dev` inicia el entorno de desarrollo.
+- `npm run build` genera la versión optimizada.
+- `npm run preview` sirve la build para revisión.
 
-If you are developing a production application, we recommend using TypeScript with type-aware lint rules enabled. Check out the [TS template](https://github.com/vitejs/vite/tree/main/packages/create-vite/template-react-ts) for information on how to integrate TypeScript and [`typescript-eslint`](https://typescript-eslint.io) in your project.
+Los modelos 3D se ubican ahora en `public/models`, lo que permite referenciarlos fácilmente desde las vistas.
