+++ pagina_apoyo/src/main.jsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import './index.css'
-import App from './App.jsx'
+import './styles/index.css'
+import App from './components/App.jsx'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
